Improve fw update robustness to boot part corruption
This splits the boot partition into two like what is currently done with
the root filesystem. Previously, renaming files on the the FAT partition
was done to make things more atomic. This got out of hand with all of
the dtb files and it wasn't too hard to conceive of ways that the FAT
filesystem could become corrupt and affect future updates. Now a clean
FAT filesystem is formatted for the boot partition, the new files added
to it, and then at the very end, the MBR update to simultaneously swap
the boot partition and the rootfs.

The downside of this approach is that if any files are stored on the
boot partition by the application, they will be lost. This sometimes is
done as a one-time manufacturing provisioning step, but is not present
in any open-source use. Anyone using this approach will need to make a
custom fwup.conf to keep the old behavior.

Interestingly enough, it's possible to update an old installation with
this new fwup configuration since the rootfs partition didn't move
locations and that's the only thing checked when figuring out A vs. B.