Restore ethernet on Raspberry Pi 1 B/B+
The Raspberry Pi B and B+ ethernet is provided by a built-in USB
Ethernet adapter using the SMSC LAN9514 chip. The driver for this can be
included via `CONFIG_USB_NET_SMSC95XX`.

This option was previously set, but appears to have been accidentally
dropped in the upgrade to Linux kernel 4.14
(b94af06ba7363de108b91d40ae7a685079f50047), resulting in the absence of
the `eth0` device on boards with this firmware version.