Start rngd early to avoid blocking the BEAM
There's a call to get_random(2) when loading the crypto NIF. Currently,
it's loaded before nerves_runtime can start rngd and provide sufficient
entropy to the kernel. This means gen_random(2) blocks the BEAM, and it
can block the it sufficiently long to trigger a hardware watchdog to
reboot. The workaround is to start rngd here.